Centralise location of atexit handlers add no-op weakref.finalize
The first atexit finalizer in the chain of finalizers that distributed
registers needs to be one in the weakref module. Otherwise we observe
hangs in the shutdown of UCX-Py where one thread has the gil but not
the UCX spinlock and vice-versa.

To ensure this, register a no-op handler at the top of `__init__.py`.

While we're here, collate all atexit handlers currently in distributed
in one place to ensure things are clear.