Minimal change to work around #7726
The first atexit finalizer in the chain of finalizers that distributed
registers needs to be one in the weakref module. Otherwise we observe
hangs in the shutdown of UCX-Py where one thread has the gil but not
the UCX spinlock and vice-versa.

To ensure this, register a no-op handler at the top of `__init__.py`.